+++ web/html/src/components/utils/HelpIcon.js
@@ -1,0 +1,14 @@
+// @flow
+import * as React from 'react';
+
+type Props = {
+  /** Title of the icon */
+  text: ?string,
+}
+
+/** Display help icon with a title */
+const HelpIcon = ({text} : Props) : React.Node => {
+    return text ? <i className="fa fa-question-circle" title={text}></i> : null;
+}
+
+export default HelpIcon;
