+++ web/html/src/components/utils/HelpIcon.stories.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import HelpIcon from './HelpIcon';
+
+export default {
+  component: HelpIcon,
+  title: 'Utils/HelpIcon',
+}
+
+export const example = () => <HelpIcon text="Title text"/>;
+
